test(aktiv-plan): del aksel modal-mock mellom tester
Ekstraherer en felles ds-react Modal-mock til testverktøyene og gjenbruker den i de to aktiv-plan-testene som tidligere hadde identisk inline-mock.

Closes #672
